mamba_ocr/pyproject.toml
[project]
name = "mamba_ocr"
version = "0.1.0"
description = """\
    This package is testing whether MAMBA is able to use in-context learning\
    to do OCR.\
    """
dependencies = [
    "numpy==1.26.3",
    "pillow==10.2.0",
    "datasets==3.1.0",
]

[project.optional-dependencies]
client = [
    "matplotlib==3.9.2",
    "PyQt6==6.7.3",
]
unity = [
    "causal-conv1d==1.4.0",
    "mamba-ssm==2.2.2",
]

[tool.distutils.egg_info]
egg_base = "."
